Added type as an option now, just like with R's plot cmd.
You can have p, b, o or l be the values of type. We use p as default. Right now
b and o are the same, but from the looks of it in R b has a little buffer between
the points and the lines connecting them. I'll tackle that later, as its not really
worth it now.