Fixed bug that ruining original option object.
Now we copy the options object before zipping it up and loading
it into the svgplot. While this requires extra space, it also
affords us to use the existing code and also preserve the original
options object. If space becomes a big issue soon, we can rethink
this to avoid making a copy.